refactor: rename story files, mocks directories
