+++ src/components/ContentNav/ContentNav.storybook.tsx
@@ -3,7 +3,7 @@ import { storiesOf } from '@storybook/react';
 import centered from '@storybook/addon-centered';
 
 import { ContentNav } from '.';
-import { itemMock } from './mocks';
+import { itemMock } from './__mocks__';
 
 storiesOf('ContentNav', module)
   .addDecorator(centered)
