+++ src/components/DesktopFilter/DesktopFilter.storybook.tsx
@@ -1,7 +1,7 @@
 import React from 'react';
 import { CSEFilterParams, CSEFilterOptions } from '../../models/filter';
 import { DesktopFilter } from '.';
-import { filterModelGS, filterModelGSC, filterModelCSCT } from '../MobileFilter/mocks';
+import { filterModelGS, filterModelGSC, filterModelCSCT } from '../MobileFilter/__mocks__';
 import { storiesOf } from '@storybook/react';
 import centered from '@storybook/addon-centered';
 import { FilterOptionModel, FilterType } from '@osu-cass/sb-components';
