+++ src/components/GenericContentPage/GenericContentPage.storybook.tsx
@@ -3,7 +3,7 @@ import { storiesOf } from '@storybook/react';
 import { GenericContentPage } from '.';
 import { MobileGenericContentPage } from './mobile';
 import { DesktopGenericContentPage } from './desktop';
-import { contentSectionsMocks } from './__tests__/mocks';
+import { contentSectionsMocks } from './__mocks__';
 import { FullPageDecorator } from '../../__decorators__';
 
 storiesOf('Generic Content Page DontTest', module)
