+++ src/components/MobileClaimTargetSummary/__mocks__/index.ts
